Moving the ReportFormatter class out of validation.py and into util.py so that other classes can use it
