+++ tiny/rna/util.py
@@ -34,6 +34,74 @@ def report_execution_time(step_name: str):
     return timer
 
 
+class ReportFormatter:
+    error_header   = "========[ ERROR ]=============================================="
+    warning_header = "========[ WARNING ]============================================"
+
+    def __init__(self, key_mapper):
+        self.key_mapper = key_mapper
+        self.warnings = []
+        self.errors = []
+
+    def print_report(self):
+        for error_section in self.errors:
+            print(self.error_header)
+            print(error_section)
+            print()
+
+        for warning_section in self.warnings:
+            print(self.warning_header)
+            print(warning_section)
+            print()
+
+    def add_error_section(self, header, body=None):
+        self.add_section(header, body, self.errors)
+
+    def add_warning_section(self, header, body=None):
+        self.add_section(header, body, self.warnings)
+
+    def add_section(self, header, body, dest):
+        if isinstance(body, dict):
+            body = self.nested_dict(body)
+        elif isinstance(body, list):
+            body = '\n'.join(body)
+        if body:
+            dest.append('\n'.join([header, body]))
+        else:
+            dest.append(header)
+
+    def nested_dict(self, summary, indent='\t'):
+        report_lines = self.recursive_indent(summary, indent)
+        return '\n'.join(report_lines)
+
+    def recursive_indent(self, mapping: dict, indent: str):
+        """Converts a nested dictionary into a properly indented list of lines
+
+        Args:
+            mapping: the nested dictionary to be formatted
+            indent: the indent string to prepend to lines on the current level
+        """
+
+        lines = []
+        for key, val in mapping.items():
+            key_header = f"{indent}{self.key_mapper.get(key, key)}: "
+            if isinstance(val, dict):
+                lines.append(key_header)
+                if len(val):
+                    lines.extend(self.recursive_indent(val, indent + '\t'))
+                else:
+                    lines.append(indent + "\t" + "{ NONE }")
+            elif isinstance(val, (list, set)):
+                lines.append(key_header)
+                if len(val):
+                    lines.extend([indent + '\t' + line for line in map(str, val)])
+                else:
+                    lines.append(indent + "\t" + "[ NONE ]")
+            else:
+                lines.append(key_header + str(val))
+        return lines
+
+
 class SmartFormatter(argparse.HelpFormatter):
     """Properly formats argparse help string fields that contain newlines.
